improve mistakes

+++ task_02/src/heap.hpp
@@ -1,5 +1,6 @@
 #pragma once
 
+#include <cstddef>
 #include <stdexcept>
 #include <vector>
 
@@ -7,8 +8,8 @@ template <typename T>
 class MinHeap {
  public:
   void Push(T value);
-  void ShiftUp(int index);
-  void ShiftDown(int index);
+  void ShiftUp(size_t index);
+  void ShiftDown(size_t index);
 
   T Pop();
   size_t Size();
@@ -40,7 +41,7 @@ size_t MinHeap<T>::Size() {
 }
 
 template <typename T>
-void MinHeap<T>::ShiftUp(int index) {
+void MinHeap<T>::ShiftUp(size_t index) {
   int parent = index / 2;
   if (index > 0 && data_[index] < data_[parent]) {
     std::swap(data_[index], data_[parent]);
@@ -50,7 +51,7 @@ void MinHeap<T>::ShiftUp(int index) {
 }
 
 template <typename T>
-void MinHeap<T>::ShiftDown(int index) {
+void MinHeap<T>::ShiftDown(size_t index) {
   int left = 2 * index + 1;
   int right = 2 * index + 2;
